refactor: circle-session-detail-view を責務ごとに分割 (#348)
1185行の巨大コンポーネントを以下の4ファイルに分割:
- match-utils.ts: ヘルパー関数群と共通型定義
- match-matrix-table.tsx: 対局結果マトリクステーブル
- match-dialog.tsx: 対局追加/編集ダイアログ
- match-delete-dialog.tsx: 削除確認ダイアログ
